refactor(sig): move sig parsing to file
