+++ src/event_message/mod.rs
@@ -65,6 +65,7 @@ impl EventMessage {
     /// Extract Serialized Data Set
     ///
     /// returns the serialized extracted data set (for signing/verification) for this event message
+    /// NOTE: this method, for deserialized events, will be UNABLE to preserve ordering
     pub fn extract_serialized_data_set(&self) -> Result<String, Error> {
         dfs_serializer::to_string(self)
     }
@@ -80,6 +81,20 @@ impl SignedEventMessage {
             event_message: message.clone(),
             signatures: sigs,
         }
+    }
+
+    pub fn serialize(&self) -> Result<Vec<u8>, Error> {
+        Ok([
+            self.event_message.serialize()?,
+            get_sig_count(self.signatures.len() as u16)
+                .as_bytes()
+                .to_vec(),
+            self.signatures
+                .iter()
+                .map(|sig| sig.to_str().as_bytes().to_vec())
+                .fold(vec![], |acc, next| [acc, next].concat()),
+        ]
+        .concat())
     }
 }
