+++ src/prefix/parse.rs
@@ -1,0 +1,75 @@
+use crate::prefix::{attached_signature::b64_to_num, AttachedSignaturePrefix, SelfSigningPrefix};
+use nom::{branch::*, combinator::*, error::ErrorKind, multi::*, sequence::*};
+
+// TODO this could be a lot nicer, but is currently written to be careful and "easy" to follow
+pub fn signature(s: &str) -> nom::IResult<&str, AttachedSignaturePrefix> {
+    let (more, type_c) = nom::bytes::complete::take(1u8)(s)?;
+
+    match type_c {
+        "A" => {
+            let (maybe_sig, index_c) = nom::bytes::complete::take(1u8)(more)?;
+
+            let index =
+                b64_to_num(index_c).map_err(|_| nom::Err::Error((index_c, ErrorKind::IsNot)))?;
+
+            let (rest, sig_s) = nom::bytes::complete::take(86u8)(maybe_sig)?;
+
+            let sig = SelfSigningPrefix::Ed25519Sha512(
+                base64::decode_config(sig_s, base64::URL_SAFE)
+                    .map_err(|_| nom::Err::Error((index_c, ErrorKind::IsNot)))?,
+            );
+
+            Ok((rest, AttachedSignaturePrefix::new(index, sig)))
+        }
+        "B" => {
+            let (maybe_sig, index_c) = nom::bytes::complete::take(1u8)(more)?;
+
+            let index =
+                b64_to_num(index_c).map_err(|_| nom::Err::Error((index_c, ErrorKind::IsNot)))?;
+
+            let (rest, sig_s) = nom::bytes::complete::take(86u8)(maybe_sig)?;
+
+            let sig = SelfSigningPrefix::ECDSAsecp256k1Sha256(
+                base64::decode_config(sig_s, base64::URL_SAFE)
+                    .map_err(|_| nom::Err::Error((index_c, ErrorKind::IsNot)))?,
+            );
+
+            Ok((rest, AttachedSignaturePrefix::new(index, sig)))
+        }
+        "0" => {
+            let (maybe_count, type_c_2) = nom::bytes::complete::take(1u8)(more)?;
+            match type_c_2 {
+                "A" => {
+                    let (maybe_sig, index_c) = nom::bytes::complete::take(2u8)(maybe_count)?;
+
+                    let index = b64_to_num(index_c)
+                        .map_err(|_| nom::Err::Error((index_c, ErrorKind::IsNot)))?;
+
+                    let (rest, sig_s) = nom::bytes::complete::take(152u8)(maybe_sig)?;
+
+                    let sig = SelfSigningPrefix::Ed448(
+                        base64::decode_config(sig_s, base64::URL_SAFE)
+                            .map_err(|_| nom::Err::Error((index_c, ErrorKind::IsNot)))?,
+                    );
+
+                    Ok((rest, AttachedSignaturePrefix::new(index, sig)))
+                }
+                _ => Err(nom::Err::Error((type_c_2, ErrorKind::IsNot))),
+            }
+        }
+        _ => Err(nom::Err::Error((type_c, ErrorKind::IsNot))),
+    }
+}
+
+#[test]
+fn test() {
+    assert_eq!(
+            signature("AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA"),
+            Ok(("", AttachedSignaturePrefix::new(0, SelfSigningPrefix::Ed25519Sha512([0u8; 64].to_vec()))))
+        );
+
+    assert_eq!(
+            signature("BCAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA"),
+            Ok(("AA", AttachedSignaturePrefix::new(2, SelfSigningPrefix::ECDSAsecp256k1Sha256([0u8; 64].to_vec()))))
+        );
+}
